requirements.txt
setuptools~=51.1.1
Pillow~=8.1.0
numba~=0.52.0